moved image files from temp folder to permanent folder
created a constant variable that is used to hold the file path string to move the image file from the temporary folder to a permanent image folder called images. The variable is used in addScore.php and index.php. Now images can be properly uploaded and displayed in the application.