updated add score and added new variable constant
added a max file size constant variable in the appVars.php file. Improved the addScore.php with statements that validate the image file. The application displays error message if the image file is exceeds 32 KB in size and if the file is not the right type (i.e. png, gif, jpeg).